Refactor feed components to be based on posts instead of event items.
Reduced some of the complexity, moved components around, etc.

Tracking post likes/unlikes now as well.
