--- src/app/components/activity/feed/event-item/blocked/blocked.ts
@@ -1,10 +1,0 @@
-import Vue from 'vue';
-import { Component, Emit, Prop } from 'vue-property-decorator';
-import { propRequired } from '../../../../../../utils/vue';
-
-@Component({})
-export default class AppActivityFeedEventItemBlocked extends Vue {
-	@Prop(propRequired(String)) username!: string;
-
-	@Emit('show') show() {}
-}
