+++ src/app/components/activity/feed/post/post.styl
